Refactor auth guard
